+++ src/components/ShoutBox.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ShoutBox = () => {
+  return (
+    <div>ShoutBox</div>
+  )
+}
+
+export default ShoutBox