+++ CHANGELOG.md
@@ -8,16 +8,23 @@ All notable changes to this project will be documented in this file.
 
 - Initial filetree and project setup
 
-## \[0.0.1\] - 2023-02-24
+## \[0.0.1\] - 2023-03-20
 
  Changed
 
 - Updated README with problem domain and user stories
 - Added App.js with basic routes set up
 
-## \[0.0.2\] - 2023-02-24
+## \[0.0.2\] - 2023-03-20
 
  Changed
 
 - Created CHANGELOG.md, and ESTIMATES.md
 - Created repository and synced to new Netlify site
+
+## \[0.0.3\] - 2023-03-24
+
+ Changed
+
+- Created Header and it's child components.
+- Added firebase and email login functionality.